requirements.txt
gradio>=4.44.0
requests>=2.31.0
feedparser>=6.0.11
python-dotenv>=1.0.1
tldextract>=5.1.2
beautifulsoup4>=4.12.2
pydantic>=2.8.2